% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_functions.R
\name{dummy_function2}
\alias{dummy_function2}
\title{A Dummy Function 2}
\usage{
dummy_function2(x)
}
\arguments{
\item{x}{A numeric value.}
}
\description{
This function calculates the square of a number.
}
