% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_functions.R
\name{dummy_function3}
\alias{dummy_function3}
\title{A Dummy Function 3}
\usage{
dummy_function3(a, b)
}
\arguments{
\item{a}{A string.}

\item{b}{Another string.}
}
\description{
This function concatenates two strings.
}
