% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_functions.R
\name{dummy_function4}
\alias{dummy_function4}
\title{A Dummy Function 4}
\usage{
dummy_function4(n)
}
\arguments{
\item{n}{A positive integer.}
}
\description{
This function generates a sequence from 1 to n.
}
