% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_functions.R
\name{dummy_function5}
\alias{dummy_function5}
\title{A Dummy Function 5}
\usage{
dummy_function5(v)
}
\arguments{
\item{v}{A numeric vector.}
}
\description{
This function returns the mean of a numeric vector.
}
